Rename the module containing utility functions and wrap methods in a class. Rewrite file name listing code in golive1.
